chore: Extract a shared eslint config
