--- packages/config-eslint/base.js
@@ -1,107 +1,0 @@
-import js from "@eslint/js";
-import globals from "globals";
-import tseslint from "typescript-eslint";
-import turboConfig from "eslint-config-turbo/flat";
-import eslintConfigPrettier from "eslint-config-prettier";
-import langfusePlugin from "@repo/eslint-plugin";
-import "eslint-plugin-only-warn";
-
-export default tseslint.config(
-  // Global ignores
-  {
-    name: "langfuse/ignores",
-    ignores: [
-      "**/node_modules/",
-      "**/dist/",
-      "**/build/",
-      "**/coverage/",
-      "**/.next/",
-      "**/.*",
-      "eslint.config.mjs",
-    ],
-  },
-
-  // Base JS rules (same as eslint v8 library.js)
-  js.configs.recommended,
-
-  // Turbo monorepo rules
-  ...turboConfig,
-
-  // Disable ESLint rules that conflict with Prettier formatting.
-  eslintConfigPrettier,
-
-  // Global settings
-  {
-    name: "langfuse/base/globals",
-    languageOptions: {
-      ecmaVersion: 2022,
-      sourceType: "module",
-      globals: {
-        ...globals.node,
-        ...globals.es2021,
-        React: "readonly",
-        JSX: "readonly",
-      },
-    },
-    rules: {
-      "no-void": "warn",
-      "no-else-return": "warn",
-      "no-unneeded-ternary": "warn",
-      "no-redeclare": "off",
-      "import/order": "off",
-    },
-  },
-
-  // TypeScript-specific - parser only + custom rules
-  // Note: Old library.js had no TS rules, only eslint:recommended
-  // Adding parser + plugin to support custom rules, but not extending recommended
-  {
-    name: "langfuse/base/typescript",
-    files: ["**/*.ts", "**/*.tsx", "**/*.mts", "**/*.cts"],
-    plugins: {
-      "@typescript-eslint": tseslint.plugin,
-    },
-    languageOptions: {
-      parser: tseslint.parser,
-      parserOptions: {
-        projectService: true,
-      },
-    },
-    rules: {
-      "no-undef": "off", // TypeScript handles this
-      "no-dupe-class-members": "off", // TypeScript handles this (and supports overloads)
-      "no-unused-vars": "off", // Use @typescript-eslint/no-unused-vars instead
-      "no-restricted-globals": [
-        "error",
-        {
-          name: "redis",
-          message: "Import redis explicitly from '@langfuse/shared/src/server'",
-        },
-      ],
-      // Custom rule from eslint v8 shared/.eslintrc.js
-      "@typescript-eslint/no-unused-vars": [
-        "warn",
-        {
-          argsIgnorePattern: "^_",
-          varsIgnorePattern: "^_",
-          caughtErrorsIgnorePattern: "^_",
-          destructuredArrayIgnorePattern: "^_",
-          ignoreRestSiblings: true,
-        },
-      ],
-      "@typescript-eslint/no-deprecated": "warn",
-      "@typescript-eslint/no-inferrable-types": "warn",
-    },
-  },
-
-  // Vitest in-source testing should only be used while developing, not in committed code.
-  {
-    name: "langfuse/no-in-source-vitest",
-    plugins: {
-      "@repo": langfusePlugin,
-    },
-    rules: {
-      "@repo/no-in-source-vitest": "warn",
-    },
-  },
-);
