+++ packages/config-eslint/index.js
@@ -1,2 +1,2 @@
 // @ts-check
-export { default } from "./base.js";
+export { default } from "./library.js";
