+++ packages/config-eslint/library.d.ts
