+++ packages/config-eslint/shared.d.ts
@@ -1,0 +1,5 @@
+import type { Linter } from "eslint";
+
+declare const config: Linter.Config[];
+
+export default config;
